Bugfixing and folder cleaning
